moved word length function to backend
colors in config now are shorts

+++ config.h
@@ -3,10 +3,10 @@
 
 static char* escape_delay = "0";
 
-static unsigned color_background = -1;
-static unsigned color_foreground = 8;
-static unsigned color_correct = 7;
-static unsigned color_incorrect = 1;
+static short color_background = -1;
+static short color_foreground = 8;
+static short color_correct = 7;
+static short color_incorrect = 1;
 
 static unsigned key_up = 107;
 static unsigned key_down = 106;
